chore: rename internal folders
